Setup base for realm
